refactor folder structure to separate out controllers
