+++ src/controllers/user/user.statics.ts
@@ -1,10 +1,9 @@
-import { IAdaptedUser } from "../utils/create-user-from-google-data";
-import { UserModel } from "./user.model";
-import { IUserDocument, IUserModel } from "./user.types";
+import { UserModel } from "../../models/user/user.model";
+import { IUser, IUserDocument, IUserModel } from "../../models/user/user.types";
 
 export async function findOneByGoogleIdOrCreate(
   this: IUserModel,
-  user: IAdaptedUser
+  user: IUser
 ): Promise<IUserDocument> {
   const userDocuments = await UserModel.find({
     "auth.googleId": user.auth.googleId,
