+++ src/controllers/user/utils/create-dummy-users.ts
@@ -1,5 +1,5 @@
-import { UserModel } from "../user.model";
-import { IUser, IUserDocument } from "../user.types";
+import { UserModel } from "../../../models/user/user.model";
+import { IUser, IUserDocument } from "../../../models/user/user.types";
 
 // const MOCK_USERS = require("../course-repo-sanitized-users.json");
 import { sanitizedUsers } from "../course-repo-sanitized-users";
