+++ src/controllers/user/utils/secure-user-adapter.test.ts
